+++ capability/itemhandler/SlotlessItemHandlerWrapper.java
@@ -156,7 +156,7 @@ public abstract class SlotlessItemHandlerWrapper implements ISlotlessItemHandler
     }
 
     @Override
-    public int getLimit() {
+    public long getLimit() {
         int total = 0;
         for (int i = 0; i < itemHandler.size(); i++) {
             total += itemHandler.getCapacityAsInt(i, itemHandler.getResource(i));
